### 2021-11-12 #3907 @whedon
[PRE REVIEW]: OpenCADD-KLIFS: A Python package to fetch kinase data from the KLIFS database
**Submitting author:** @dominiquesydow (<a href="http://orcid.org/0000-0003-4205-8705">Dominique Sydow</a>)
**Repository:** <a href="https://github.com/volkamerlab/opencadd" target ="_blank">https://github.com/volkamerlab/opencadd</a>
**Version:** v1.0.0
