### 2021-11-12 #3907 @kyleniemeyer
Hi @dominiquesydow, it looks like your JOSS paper is not located in the repository correctly for our system to find. I think this may be because the paper is nested? Normally the paper is located in `paper/paper.md`.

@openjournals/dev can you give more-specific guidance here?
