### 2021-11-12 #3907 @danielskatz
@kyleniemeyer - I don't see a paper at all, in the tree or in a branch - JOSS should find it if it's in a directory. So maybe before an editor starts, we should make sure there is a paper added?
