### 2021-11-12 #3907 @kyleniemeyer
@danielskatz the paper appears to be here: https://github.com/volkamerlab/opencadd/tree/master/papers/opencadd-klifs
