### 2021-11-13 #3907 @dominiquesydow
Hi @kyleniemeyer and @danielskatz,

Thank you for opening this pre-review!

I am sorry for the holdup caused by the nested folder structure; I assumed it will work, since the [Whedon paper review service](https://whedon.theoj.org/) can locate the paper:
- Paper repository address: https://github.com/volkamerlab/opencadd
- Custom branch (optional): master

---

> @danielskatz the paper appears to be here: https://github.com/volkamerlab/opencadd/tree/master/papers/opencadd-klifs

That is correct.

I have moved the files to an un-nested `paper` folder on the `joss-paper` branch; https://github.com/volkamerlab/opencadd/pull/121.
I suggest I merge that PR and we try again?

> I see - whedon usually finds things like this - I wonder if the directory was called joss if that would help

Alternatively, please let me know if you prefer that I try `paper/joss` (if I understand your suggestion correctly).

---

As a general question: Are published JOSS papers linked to a specific `paper.md` location or can authors move these files after publication to e.g. subdirectories?
