### 2021-11-13 #3907 @kyleniemeyer
@dominiquesydow Hmm, I'm really not sure what is going on. I can see the paper in `paper/paper.md` in the `joss-paper` branch, which should totally work. @openjournals/dev @arfon any ideas what is going on here?

Also, you can absolutely leave the paper in a branch that is never merged, or remove the paper after publication. We archive the paper itself and the software repository does not need to contain it after acceptance.
