### 2021-11-16 #3907 @arfon
@dominiquesydow – apologies for the issues here. I've found the original issue (the repository address had a space in it which was confusing Whedon).

The paper currently won't compile as there are two `paper.md` files present. (`/papers/opencadd-klifs/paper.md`, `/paper/paper.md`). Please delete one of these (either is fine).
