### 2021-11-16 993a032
Delete duplicated OpenCADD-KLIFS paper
