+++ papers/opencadd-klifs/paper.bib
@@ -1,152 +1,0 @@
-@article{Cohen:2021,
-    title={Kinase drug discovery 20 years after imatinib: progress and future directions},
-    author={Cohen, Philip and Cross, Darren and J{\"a}nne, Pasi A.},
-    journal={Nature Reviews Drug Discovery},
-    volume={20},
-    number={7},
-    pages={551-569},
-    year={2021},
-    doi={10.1038/s41573-021-00195-4},
-}
-
-@article{Kooistra:2017,
-    author = {Kooistra, Albert J. and Volkamer, Andrea},
-    title = {{Kinase-Centric Computational Drug Development}},
-    journal = {Annu. Rep. Med. Chem.},
-    volume = {50},
-    pages = {197--236},
-    year = {2017},
-    doi = {10.1016/BS.ARMC.2017.08.001},
-}
-
-@article{Kanev:2021,
-    title = "{KLIFS: an overhaul after the first 5 years of supporting kinase research}",
-    author = {Kanev, Georgi K and {de Graaf}, Chris and Westerman, Bart A and {de Esch}, Iwan J P and Kooistra, Albert J},
-    journal = {Nucleic Acids Research},
-    volume = {49},
-    number = {D1},
-    pages = {D562-D569},
-    year = {2020},
-    doi = {10.1093/nar/gkaa895},
-}
-
-@article{vanLinden:2014,
-    author={van Linden, Oscar P. J. and Kooistra, Albert J. and Leurs, Rob and de Esch, Iwan J. P. and de Graaf, Chris},
-    title={KLIFS: A Knowledge-Based Structural Database To Navigate Kinase--Ligand Interaction Space},
-    journal={Journal of Medicinal Chemistry},
-    volume={57},
-    number={2},
-    pages={249-277},
-    year={2014},
-    doi={10.1021/jm400378w},
-}
-
-@article{Raschka:2017,
-    title = {BioPandas: Working with molecular structures in pandas DataFrames},
-    author = {Sebastian Raschka},
-    journal = {The Journal of Open Source Software},
-    volume = {2},
-    number = {14},
-    year  = {2017},
-    doi = {10.21105/joss.00279},
-}
-
-@inproceedings{Kluyver:2016,
-    booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
-    editor = {Fernando Loizides and Birgit Scmidt},
-    title = {Jupyter Notebooks - a publishing format for reproducible computational workflows},
-    author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and  Jupyter development team},
-    publisher = {IOS Press},
-    year = {2016},
-    pages = {87--90},
-    url = {https://eprints.soton.ac.uk/403913/},
-}
-
-@misc{klifsswagger,
-    author = {KLIFS},
-    title = {{KLIFS OpenAPI}},
-    year = 2021,
-    publisher = {https://dev.klifs.net},
-    url = {https://dev.klifs.net/swagger_v2/},
-}
-
-@misc{bravado,
-    author = {bravado},
-    title = {{bravado}},
-    year = 2021,
-    publisher = {GitHub},
-    journal = {GitHub repository},
-    url = {https://github.com/Yelp/bravado},
-}
-
-@misc{pandas,
-    author = {{The pandas development team}},
-    title = {pandas-dev/pandas: Pandas},
-    year = 2020,
-    publisher = {Zenodo},
[…]
