### 2021-11-17 f889208
Correct docs for KLIFS RMSD values
