### 2021-11-24 #3951 @whedon
[REVIEW]: OpenCADD-KLIFS: A Python package to fetch kinase data from the KLIFS database
**Submitting author:** @dominiquesydow (<a href="http://orcid.org/0000-0003-4205-8705">Dominique Sydow</a>)
**Repository:** <a href="https://github.com/volkamerlab/opencadd" target ="_blank">https://github.com/volkamerlab/opencadd</a>
**Version:** v1.0.1

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href="https://github.com/volkamerlab/opencadd " target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

✨ Important: Please do not use the _Convert to issue_ functionality when working through this checklist, instead, please open any new issues associated with your review <a target="_blank" href="https://github.com/volkamerlab/opencadd ">in the software repository associated with the submission</a>. ✨

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/volkamerlab/opencadd ">repository url</a>?

✨ Important: Please do not use the _Convert to issue_ functionality when working through this checklist, instead, please open any new issues associated with your review <a target="_blank" href="https://github.com/volkamerlab/opencadd ">in the software repository associated with the submission</a>. ✨

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/volkamerlab/opencadd ">repository url</a>?

✨ Important: Please do not use the _Convert to issue_ functionality when working through this checklist, instead, please open any new issues associated with your review <a target="_blank" href="https://github.com/volkamerlab/opencadd ">in the software repository associated with the submission</a>. ✨

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/volkamerlab/opencadd ">repository url</a>?
