### 2021-11-30 4a657bc
OpenCADD-KLIFS: Do not cite reference twice
