### 2021-11-30 volkamerlab/opencadd#126 @mcs07
Installation docs assume conda-forge channel already in conda config
It's really great that you have a package on `conda-forge` for easy installation!

This might be a bit of an edge case, but if users install `mamba` into an existing conda install (e.g. from Miniconda or Anaconda), they won't automatically get the `conda-forge` channel pre-configured (like it would be from a `mambaforge` install). In this case, the installation instructions as written won't work.

It could be worth extending the existing "Note" about mamba to mention this, for example:

> If you installed `mamba` into an existing `conda` installation, also make sure that the `conda-forge` channel is configured by running `conda config --add channels conda-forge`.

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/3951
