### 2021-12-06 volkamerlab/opencadd#128 @andrewtarzia
Get PDF compiled for JOSS review
As stated in title - it would be good to read this, rather than the MD file.

Link to JOSS review: openjournals/joss-reviews#3951
