### 2021-12-07 da0c90d
Docs: Add cf channel if mamba was installed into existing conda env
