### 2021-12-07 volkamerlab/opencadd#126 @dominiquesydow
Hi @mcs07,

Thanks a lot for this suggestion! I have added the suggested note to our documentation.
Would you like to take a look before I merge?
https://github.com/volkamerlab/opencadd/pull/130
