### 2021-12-07 #3907 @dominiquesydow
> @dominiquesydow – apologies for the issues here. I've found the original issue (the repository address had a space in it which was confusing Whedon).
>
> The paper currently won't compile as there are two `paper.md` files present. (`/papers/opencadd-klifs/paper.md`, `/paper/paper.md`). Please delete one of these (either is fine).

Hi @afron, my apologies for circling back to the previous Whedon issue. We may have a same problem again in the review PR:
https://github.com/openjournals/joss-reviews/issues/3951

We have now on `opencadd`'s master the following files (FYI the merged PR https://github.com/volkamerlab/opencadd/pull/121):
```
├── paper
│   ├── opencadd_klifs_toc.png
│   ├── paper.bib
│   └── paper.md
├── papers
│   └── opencadd-superposition
│       └── paper-superposer.md
```
Generating the PDF from the `master` branch does not work (see the fail in the review PR [here](https://github.com/openjournals/joss-reviews/issues/3951#issuecomment-987850134)).

We have only one `paper.md` file  - is this again the URL-space problem?
