### 2021-12-07 volkamerlab/opencadd#128 @dominiquesydow
Hi @andrewtarzia,

My apologies for not attending to this earlier.

We got the compilation running once in the pre-review (see https://github.com/openjournals/joss-reviews/issues/3907#issuecomment-970107477).

I have asked for help to figure out why it is not compiling anymore (see https://github.com/openjournals/joss-reviews/issues/3907#issuecomment-987858595). I will get back to you once the compilation is working again.
