### 2021-12-07 9a15563
Repo installation: Offer both test env and user env
