### 2021-12-07 7978d5d
Delete unused `requirements.txt` file
