+++ devtools/conda-envs/requirements.txt
@@ -1,2 +1,0 @@
-biotite
-MDAnalysis
