### 2021-12-08 volkamerlab/opencadd#128 @dominiquesydow
The PDF compilation is working now :) See https://github.com/openjournals/joss-reviews/issues/3951#issuecomment-988790990.
