### 2021-12-08 942c1db
Update devtools/conda-envs/user_env.yaml
Co-authored-by: Estefania Barreto-Ojeda <estefania.barretooje@ucalgary.ca>
