+++ devtools/conda-envs/user_env.yaml
@@ -1,4 +1,4 @@
-name: test
+name: opencadd
 channels:
   - conda-forge
   - defaults
