### 2021-12-08 c470342
Update docs/installing.rst
Co-authored-by: Estefania Barreto-Ojeda <estefania.barretooje@ucalgary.ca>
