### 2021-12-08 cc746fa
Add new installation section for developers (+testing/docs instructions)
