### 2021-12-08 d5b22bd
Sync package order in env files with conda recipe (instead of PR #120)
