### 2021-12-08 volkamerlab/opencadd#133 @andrewtarzia
Comments on JOSS manuscript
Hello. Just read through the manuscript and a few comments:

- [x] Incomplete reference to Raschka on line 46
- [x] The manuscript does not discuss how people would have accessed this data previously. Was the Rest API not accessible through Python already? Therefore, the statement of need is also unclear as it is written (I definitely see the need for the software!).
- [x] The Statement of Need section is too long, the manuscript should have a Description or Use case section.

For https://github.com/openjournals/joss-reviews/issues/3951
