### 2021-12-09 a00996d
Update devtools/conda-envs/user_env.yaml
Co-authored-by: Estefania Barreto-Ojeda <estefania.barretooje@ucalgary.ca>
