+++ devtools/conda-envs/user_env.yaml
@@ -10,7 +10,7 @@ dependencies:
   - numpy
   - pandas
   - matplotlib-base
-  - mdanalysis
+  - MDAnalysis
   - nglview
   # Remove jsonschema once this issue is fixed: https://github.com/Yelp/bravado/issues/478
   - jsonschema<4.0.0
