### 2021-12-09 8f614a8
Remove numpy and jsonschema from envs
