+++ devtools/conda-envs/user_env.yaml
@@ -7,13 +7,10 @@ dependencies:
   # Base depends
   - python
   - pip
-  - numpy
   - pandas
   - matplotlib-base
   - MDAnalysis
   - nglview
-  # Remove jsonschema once this issue is fixed: https://github.com/Yelp/bravado/issues/478
-  - jsonschema<4.0.0
   - bravado
   - biopandas
   - rdkit
