### 2021-12-09 255699a
Remove numpy/jsonschema from docs depends
