+++ docs/installing.rst
@@ -90,11 +90,9 @@ Dependencies
 - ``pandas``
 - ``biopandas`` [1,2]
 - ``bravado`` [1]
-- ``jsonschema`` <4.0.0 [1] (pinned until `this issue <https://github.com/Yelp/bravado/issues/478>`_ is resolved)
 - ``rdkit`` [1]
 - ``tqdm`` [1]
 - ``biopython`` <=1.77 [2]
-- ``numpy`` [3]
 - ``mdanalysis`` (with support of 2.0) [3]
 - ``biotite`` [3]
 - ``mmligner`` [3]
