### 2021-12-09 #3951 @mcs07
To update: I am currently testing the functionality. General checks, installation, documentation - all looks good. I have added some additional comments and suggestions about the manuscript in https://github.com/volkamerlab/opencadd/issues/133.
