### 2021-12-09 volkamerlab/opencadd#134 @ojeda-e
State of the field missing in JOSS paper 
The JOSS paper should include a comparison of openCADD-KLIFS to other commonly-used packages. Here are some questions that may help to add a couple of lines to the state of the field in your paper:
- Are there any other similar packages to openCADD-KLIFS? If yes, how is openCADD-KLIFS different?
- Is this package the open-source version of proprietary software?
- Is openCADD-KLIFS offering a unique solution to fetch data? Is openCADD-KLIFS the only automated package to fetch KLIFS data?
- How was the process of fetching KLIFS data before openCADD? How is the package making this process different/better/faster?

To fix this issue:

- [ ] Add a state of the field in the paper, where a comparison to other packages is offered to the reader.

REF: https://github.com/openjournals/joss-reviews/issues/3951
