### 2021-12-12 df8c302
add Clustal Omega as alignment option for MDA
