+++ opencadd/structure/superposition/benchmark.py
@@ -1,4 +1,0 @@
-import pandas
-
-
-with open 
