+++ opencadd/tests/structure/test_cli.py
@@ -1,8 +1,0 @@
-"""
-Tests for opencadd.structure.superposition.cli
-"""
-
-import pytest
-
-from opencadd.structure.core import Structure
-from opencadd.structure.superposition.engines.theseus import TheseusAligner
