+++ opencadd/structure/superposition/engines/2bbm_1cfc.txt
@@ -1,236 +1,0 @@
-[0;37;1;44mMMLigner (v1.0.2): Minimum Message Length (MML) alignment of protein structures [0;0m
-[0;1;37;1;46m   Authors:[0;1;37;1;46m    Collier, Allison, Lesk, Stuckey, Garcia de la Banda, Konagurthu. [0m
-[1;37;1;46m   Reference:[0;1;37;1;46m  Bioinformatics. 33(7):1005-1013, 2017.                           [0m
-
-Parsing Brookhaven PDB      ...structure1.pdb      nResi = 174   [ [32;4mOK[0m ]
-Parsing Brookhaven PDB      ...structure2.pdb      nResi = 148   [ [32;4mOK[0m ]
-Computing MMLigner alignment...
-...building MFP Library                                          [ [32;4mOK[0m ]
-...filtering MFP library                                                  21.925  %         24.006  %         27.362  %         37.987  %         39.324  %         53.542  %         70.483  %         76.319  %         87.853  %         100.000 %[ [32;4mOK[0m ]
-...clustering filtered MFP library                               [ [32;4mOK[0m ]
-...generating seed alignments                                             100.000 %         2.151   %         8.602   %         15.054  %         49.462  %         51.613  %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %[ [32;4mOK[0m ] 5 seeds
-...expectation maximization on seed alignments using I-value...
-   ...perturbing seed alignment 1 of 5            0.000   %         2.000   %         4.000   %         6.000   %         8.000   %         10.000  %         12.000  %         14.000  %         16.000  %         18.000  %         20.000  %         22.000  %         24.000  %         26.000  %         28.000  %         30.000  %         32.000  %         34.000  %         36.000  %         100.000 %      [ [32;4mOK[0m ]
[…]
