+++ mmligner_issue_56.txt
@@ -1,69 +1,0 @@
-[0;37;1;44mMMLigner (v1.0.2): Minimum Message Length (MML) alignment of protein structures [0;0m
-[0;1;37;1;46m   Authors:[0;1;37;1;46m    Collier, Allison, Lesk, Stuckey, Garcia de la Banda, Konagurthu. [0m
-[1;37;1;46m   Reference:[0;1;37;1;46m  Bioinformatics. 33(7):1005-1013, 2017.                           [0m
-
-Parsing Brookhaven PDB      ...structure1.pdb      nResi = 289   [ [32;4mOK[0m ]
-Parsing Brookhaven PDB      ...structure2.pdb      nResi = 276   [ [32;4mOK[0m ]
-Computing MMLigner alignment...
-...building MFP Library                                          [ [32;4mOK[0m ]
-...filtering MFP library                                                  0.822   %         1.595   %         1.595   %         1.618   %         1.631   %         1.799   %         1.824   %         3.509   %         4.948   %         5.334   %         5.951   %         5.984   %         6.065   %         6.073   %         6.403   %         6.422   %         6.475   %         7.366   %         9.161   %         14.600  %         14.636  %         16.351  %         16.358  %         16.368  %         16.566  %         17.223  %         20.266  %         20.266  %         20.458  %         20.474  %         20.551  %         20.569  %         20.717  %         26.828  %         27.275  %         27.276  %         27.510  %         28.147  %         29.923  %         29.927  %         29.998  %         30.686  %         31.082  %         31.562  %         32.572  %         33.204  %         33.946  %         37.216  %         37.238  %         37.271  %         37.305  %         37.632  %         38.082  %         38.193  %         38.576  %         39.035  %         39.644  %         39.666  %         39.680  %         39.837  %         39.840  %         40.211  %         41.148  %         41.173  %         42.934  %         46.846  %         47.550  %         48.113  %         48.138  %         48.148  %         49.044  %         49.578  %         50.752  %         50.777  %         50.993  %         51.004  %         51.111  %         52.547  %         52.596  %         53.141  %         55.518  %         57.513  %         57.705  %         57.714  %         57.721  %         61.365  %         62.784  %         63.696  %         63.706  %         63.741  %         63.745  %         63.785  %         65.872  %         65.969  %         66.084  %         66.259  %         66.751  %         66.941  %         67.847  %         68.828  %         69.403  %         69.440  %         71.185  %         72.324  %         72.767  %         72.818  %         74.034  %         74.217  %         74.805  %         75.916  %         78.925  %         81.874  %         81.935  %         88.387  %         88.702  %         89.951  %         90.614  %         91.078  %         92.969  %         93.906  %         94.229  %         95.467  %         100.000 %[ [32;4mOK[0m ]
-...clustering filtered MFP library                               [ [32;4mOK[0m ]
[…]
