+++ opencadd/structure/superposition/engines/debug.txt
@@ -1,39 +1,0 @@
-[0;37;1;44mMMLigner (v1.0.2): Minimum Message Length (MML) alignment of protein structures [0;0m
-[0;1;37;1;46m   Authors:[0;1;37;1;46m    Collier, Allison, Lesk, Stuckey, Garcia de la Banda, Konagurthu. [0m
-[1;37;1;46m   Reference:[0;1;37;1;46m  Bioinformatics. 33(7):1005-1013, 2017.                           [0m
-
-Parsing Brookhaven PDB      ...structure1.pdb      nResi = 235   [ [32;4mOK[0m ]
-Parsing Brookhaven PDB      ...structure2.pdb      nResi = 306   [ [32;4mOK[0m ]
-Computing MMLigner alignment...
-...building MFP Library                                          [ [32;4mOK[0m ]
-...filtering MFP library                                                  6.294   %         14.629  %         16.047  %         18.724  %         19.055  %         22.113  %         32.162  %         54.789  %         55.099  %         60.845  %         68.218  %         77.591  %         82.822  %         87.594  %         100.000 %[ [32;4mOK[0m ]
-...clustering filtered MFP library                               [ [32;4mOK[0m ]
-...generating seed alignments                                             100.000 %         100.000 %         100.000 %         42.105  %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %         100.000 %[ [32;4mOK[0m ] 0 seed
-...expectation maximization on seed alignments using I-value...
-   [none of the resultant alignments are significant]
-Writing aligment output to  [33;4mtemp__1.afasta[0m [ [32;4mOK[0m ]
-Writing alignment stats to  [33;4mtemp__1.stats[0m  [ [32;4mOK[0m ]
-Writing superposed pdb to   [33;4mp_superposed__1.pdb[0m       [ [32;4mOK[0m ]
-Summary:
-Alignment 1 stats
-Structure Sizes =   306, 235
-Coverage        =   0        
-RMSD            =     0.000 A
-I(A)            =    35.803 bits
-NULL(S & T)     = 18839.720 bits
-I(A & <S,T>)    = 18852.803 bits
-Compression     =   -13.083 bits (+ve better; -ve worse)
--------
-[ [32;4m ALL DONE!!![0m ]
-[36;4m.                                    -=o0o=-                                   .[0m
-[36;3m|                                   REFERENCE                                  |[0m
-[33;3m| J.Collier, L.Allison, A.Lesk, P.Stuckey, M.Garcia de la Banda, A.Konagurthu  |
-| Statistical inference of protein structural alignments using information and |
-| compression, Bioinformatics. 33(7):1005-1013, 2017.                          |[0m
-[35;3m| DOI: https://doi.org/10.1093/bioinformatics/btw757                           |[0m
-[36;4m|                                    -=o0o=-                                   |[0m
-[37;2m| See other LCB structural bioinformatics tools:                               |[0m
-[37;2m|  * MUSTANG: http://lcb.infotech.monash.edu.au/mustang                        |[0m
-[37;2m|  * SST    : http://lcb.infotech.monash.edu.au/sstweb                         |[0m
-[37;2m|  * SUPER  : http://lcb.infotech.monash.edu.au/super                          |[0m
-[37;2m|-------------------------------------0o.o0------------------------------------|[0m
