### 2021-12-13 c5aa4c0
add redo to fetch structure
