+++ devtools/conda-envs/test_env.yaml
@@ -11,6 +11,7 @@ dependencies:
   - nglview
   - pandas
   - matplotlib-base
+  - redo
   # Remove jsonschema once this issue is fixed: https://github.com/Yelp/bravado/issues/478
   - jsonschema<4.0.0
   - bravado
