+++ paper/paper.bib
@@ -62,12 +62,43 @@
     url = {https://eprints.soton.ac.uk/403913/},
 }
 
+@article{Mendez:2018,
+    author = {Mendez, David and Gaulton, Anna and Bento, A Patrícia and Chambers, Jon and De Veij, Marleen and Félix, Eloy and Magariños, María Paula and Mosquera, Juan F and Mutowo, Prudence and Nowotka, Michał and Gordillo-Marañón, María and Hunter, Fiona and Junco, Laura and Mugumbate, Grace and Rodriguez-Lopez, Milagros and Atkinson, Francis and Bosc, Nicolas and Radoux, Chris J and Segura-Cabrera, Aldo and Hersey, Anne and Leach, Andrew R},
+    title = "{ChEMBL: towards direct deposition of bioassay data}",
+    journal = {Nucleic Acids Research},
+    volume = {47},
+    number = {D1},
+    pages = {D930-D940},
+    year = {2018},
+    doi = {10.1093/nar/gky1075},
+}
+
+@article{Carles:2018,
+    author = {Carles, Fabrice and Bourg, St{\'{e}}phane and Meyer, Christophe and Bonnet, Pascal},
+    title = {{PKIDB: A Curated, Annotated and Updated Database of Protein Kinase Inhibitors in Clinical Trials}},
+    journal = {Molecules},
+    volume = {23},
+    number = {4},
+    pages = {908},
+    year = {2018},
+    doi = {10.3390/molecules23040908},
+}
+
 @misc{klifsswagger,
     author = {KLIFS},
     title = {{KLIFS OpenAPI}},
     year = 2021,
     publisher = {https://dev.klifs.net},
     url = {https://dev.klifs.net/swagger_v2/},
+}
+
+@misc{requests,
+    author = {requests},
+    title = {{requests}},
+    year = 2021,
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/psf/requests},
 }
 
 @misc{bravado,
