+++ paper/paper.bib
@@ -84,6 +84,28 @@
     doi = {10.3390/molecules23040908},
 }
 
+@article{McGuire:2017,
+    author = {McGuire, Ross and Verhoeven, Stefan and Vass, Márton and Vriend, Gerrit and de Esch, Iwan J. P. and Lusher, Scott J. and Leurs, Rob and Ridder, Lars and Kooistra, Albert J. and Ritschel, Tina and de Graaf, Chris},
+    title = {3D-e-Chem-VM: Structural Cheminformatics Research Infrastructure in a Freely Available Virtual Machine},
+    journal = {Journal of Chemical Information and Modeling},
+    volume = {57},
+    number = {2},
+    pages = {115-121},
+    year = {2017},
+    doi = {10.1021/acs.jcim.6b00686},
+}
+
+@article{Kooistra:2018,
+   author = {Kooistra, A. J.  and Vass, M.  and McGuire, R.  and Leurs, R.  and de Esch, I. J. P.  and Vriend, G.  and Verhoeven, S.  and de Graaf, C. },
+   title = {{3{D}-e-{C}hem: {S}tructural {C}heminformatics {W}orkflows for {C}omputer-{A}ided {D}rug {D}iscovery}},
+   journal = {ChemMedChem},
+   volume = {13},
+   number = {6},
+   pages = {614--626},
+   year = {2018},
+   doi = {10.1002/cmdc.201700754},
+}
+
 @misc{klifsswagger,
     author = {KLIFS},
     title = {{KLIFS OpenAPI}},
