+++ paper/paper.bib
@@ -9,14 +9,28 @@
     doi={10.1038/s41573-021-00195-4},
 }
 
-@article{Kooistra:2017,
-    author = {Kooistra, Albert J. and Volkamer, Andrea},
-    title = {{Kinase-Centric Computational Drug Development}},
-    journal = {Annu. Rep. Med. Chem.},
+@incollection{Kooistra:2017,
+    booktitle = {Platform Technologies in Drug Discovery and Validation},
+    series = {Annual Reports in Medicinal Chemistry},
+    editor = {Robert A. Goodnow},
+    title = {Chapter Six - Kinase-Centric Computational Drug Development},
+    author = {Albert J. Kooistra and Andrea Volkamer},
+    publisher = {Academic Press},
     volume = {50},
-    pages = {197--236},
+    pages = {197-236},
     year = {2017},
-    doi = {10.1016/BS.ARMC.2017.08.001},
+    doi = {10.1016/bs.armc.2017.08.001},
+}
+
+@inproceedings{Kluyver:2016,
+    booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+    editor = {Fernando Loizides and Birgit Scmidt},
+    title = {Jupyter Notebooks - a publishing format for reproducible computational workflows},
+    author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and  Jupyter development team},
+    publisher = {IOS Press},
+    year = {2016},
+    pages = {87--90},
+    url = {https://eprints.soton.ac.uk/403913/},
 }
 
 @article{Kanev:2021,
@@ -49,17 +63,6 @@
     number = {14},
     year  = {2017},
     doi = {10.21105/joss.00279},
-}
-
-@inproceedings{Kluyver:2016,
-    booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
-    editor = {Fernando Loizides and Birgit Scmidt},
-    title = {Jupyter Notebooks - a publishing format for reproducible computational workflows},
-    author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and  Jupyter development team},
-    publisher = {IOS Press},
-    year = {2016},
-    pages = {87--90},
-    url = {https://eprints.soton.ac.uk/403913/},
 }
 
 @article{Mendez:2018,
@@ -96,7 +99,7 @@
 }
 
 @article{Kooistra:2018,
-   author = {Kooistra, A. J.  and Vass, M.  and McGuire, R.  and Leurs, R.  and de Esch, I. J. P.  and Vriend, G.  and Verhoeven, S.  and de Graaf, C. },
+   author = {Kooistra, A. J. and Vass, M.  and McGuire, R. and Leurs, R. and de Esch, I. J. P. and Vriend, G. and Verhoeven, S. and de Graaf, C. },
    title = {{3{D}-e-{C}hem: {S}tructural {C}heminformatics {W}orkflows for {C}omputer-{A}ided {D}rug {D}iscovery}},
    journal = {ChemMedChem},
    volume = {13},
