### 2021-12-16 d876024
Add table of minimal supported versions
