+++ docs/installing.rst
@@ -85,24 +85,42 @@ and leave a note in the
 Dependencies
 ------------
 
-``opencadd`` requires at least Python 3.6 and needs the following modules: 
+``opencadd`` is supported for at least Python 3.7 and needs the following modules: 
 
-- ``pandas``
-- ``biopandas`` [1,2]
-- ``bravado`` [1]
-- ``rdkit`` [1]
-- ``tqdm`` [1]
-- ``biopython`` <=1.77 [2]
-- ``mdanalysis`` (with support of 2.0) [3]
-- ``biotite`` [3]
-- ``mmligner`` [3]
-- ``muscle`` [3]
-- ``theseus`` [3]
-- ``matplotlib-base`` [4]
-- ``nglview`` [4]
++---------------------+---------------------------+--------------------+--+--+
+| Package             | Minimal supported version | Used by submodules |  |  |
++=====================+===========================+====================+==+==+
+| ``pandas``          | 1.3.4                     | all                |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``biopandas``       | 0.2.9                     | 1, 2               |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``bravado``         | 11.0.3                    | 1                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``rdkit``           | 2021.09.2                 | 1                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``tqdm``            | 4.62.3                    | 1                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``biopython``       | = 1.77                    | 2                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``mdanalysis``      | 2.0.0                     | 3                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``biotite``         | 0.31.0                    | 3                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``mmligner``        | 1.0.2                     | 3                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``muscle``          | 3.8.1551                  | 3                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``theseus``         | 3.3.0                     | 3                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``matplotlib-base`` | 3.5.0                     | 4                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
+| ``nglview``         | 3.0.3                     | 4                  |  |  |
++---------------------+---------------------------+--------------------+--+--+
 
 
 Some packages are only needed for a subset of the following modules: [1] ``opencadd.databases.klifs``, 
 [2] ``opencadd.io``, 
 [3] ``opencadd.structure.superposition``, 
 [4] ``opencadd.structure.pocket``
+
+This list of minimal supported version is based on `this CI run <https://github.com/volkamerlab/opencadd/runs/4462667598?check_suite_focus=true#step:6:42>`_.
