### 2021-12-16 d34a1bd
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
