### 2021-12-16 adc42db
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
