### 2021-12-16 86db200
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
