### 2021-12-16 98b7773
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
