### 2021-12-16 4cfb671
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
