### 2021-12-16 c6659a1
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
