### 2021-12-16 196094d
Update paper/paper.md
Co-authored-by: AndreaVolkamer <andrea.volkamer@charite.de>
