### 2021-12-16 7a092bb
Implement more of Andrea's text edits
