+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
   - name: Jaime Rodríguez-Guerra
     orcid: 0000-0001-8974-1566
     affiliation: 1
-  - name: Andrea Volkamer
+  - name: Andrea Volkamer^[corresponding author]
     affiliation: 1
     orcid: 0000-0002-3760-580X
 affiliations:
