+++ paper/paper.bib
@@ -14,7 +14,7 @@
     series = {Annual Reports in Medicinal Chemistry},
     editor = {Robert A. Goodnow},
     title = {Chapter Six - Kinase-Centric Computational Drug Development},
-    author = {Albert J. Kooistra and Andrea Volkamer},
+    author = {Kooistra, {A. J.} and Volkamer, A.},
     publisher = {Academic Press},
     volume = {50},
     pages = {197-236},
@@ -45,7 +45,7 @@
 }
 
 @article{vanLinden:2014,
-    author={van Linden, Oscar P. J. and Kooistra, Albert J. and Leurs, Rob and de Esch, Iwan J. P. and de Graaf, Chris},
+    author={{van Linden}, Oscar P. J. and Kooistra, Albert J. and Leurs, Rob and de Esch, Iwan J. P. and de Graaf, Chris},
     title={KLIFS: A Knowledge-Based Structural Database To Navigate Kinase--Ligand Interaction Space},
     journal={Journal of Medicinal Chemistry},
     volume={57},
@@ -99,7 +99,7 @@
 }
 
 @article{Kooistra:2018,
-   author = {Kooistra, A. J. and Vass, M.  and McGuire, R. and Leurs, R. and de Esch, I. J. P. and Vriend, G. and Verhoeven, S. and de Graaf, C. },
+   author = {Kooistra, {A. J.} and Vass, M.  and McGuire, R. and Leurs, R. and de Esch, I. J. P. and Vriend, G. and Verhoeven, S. and de Graaf, C. },
    title = {{3{D}-e-{C}hem: {S}tructural {C}heminformatics {W}orkflows for {C}omputer-{A}ided {D}rug {D}iscovery}},
    journal = {ChemMedChem},
    volume = {13},
