### 2021-12-16 volkamerlab/opencadd#133 @dominiquesydow
Hi @mcs07 and @andrewtarzia,

Thanks again for the points you raised in this issue.
Please have a look at the revised version of the manuscript:
https://github.com/openjournals/joss-reviews/issues/3951#issuecomment-996179098

I have merged your comments with those from @ojeda-e in issue #134 in PR #135.
