### 2021-12-16 volkamerlab/opencadd#134 @dominiquesydow
Hi @ojeda-e,

Thanks again for your input on the manuscript!
Please have a look at the revised version:
https://github.com/openjournals/joss-reviews/issues/3951#issuecomment-996179098

I have merged your comments with those from issue #133 in PR #135.
