+++ docs/databases_klifs_statement_of_need.rst
@@ -1,24 +1,25 @@
 Statement of need
 ================= 
 
-OpenCADD-KLIFS is aimed at current and future users of the KLIFS database who seek to 
-integrate kinase resources into Python-based research projects.
-This module offers access to KLIFS data [Kanev_2021]_ such as information about kinases, 
-structures, ligands, 
+The KLIFS resource [Kanev_2021]_ contains information about kinases, structures, ligands, 
 interaction fingerprints, and bioactivities. 
 KLIFS thereby focuses especially on the ATP binding site, defined as a set of 85 residues and 
-aligned across all structures using a multiple sequence alignment (MSA) [vanLinden_2014]_.
-With OpenCADD-KLIFS, KLIFS data can be queried either locally from a KLIFS download or remotely 
-from the KLIFS webserver. 
-The presented module provides identical APIs for the remote and local queries for KLIFS data and 
-streamlines all output into 
-standardized `Pandas <https://doi.org/10.5281/zenodo.5574486>`_ DataFrames to allow for easy and 
-quick downstream data analyses (Figure 1). This Pandas-focused setup is ideal to work with in 
-Jupyter notebooks [Kluyver_2016]_. 
+aligned across all structures using a multiple sequence alignment [vanLinden_2014]_.
+Fetching, filtering, and integrating the KLIFS content on a larger scale into Python-based 
+pipelines is currently not straight-forward, especially for users without a background in 
+online queries. 
+Furthermore, switching between data queries from a *local* KLIFS download and 
+the *remote* KLIFS database is not readily possible.
 
-`OpenCADD-KLIFS <https://opencadd.readthedocs.io/en/latest/databases_klifs.html>`_  
-(``opencadd.databases.klifs``) is a part of the `OpenCADD <https://opencadd.readthedocs.io/>`_ 
-package, a collection of Python modules for structural cheminformatics. 
+OpenCADD-KLIFS is aimed at current and future users of the KLIFS database who seek to 
+integrate kinase resources into Python-based research projects.
+With OpenCADD-KLIFS, KLIFS data can be queried either locally from a KLIFS download or 
+remotely from the KLIFS webserver. 
+The presented module provides identical APIs for the remote and local queries and 
+streamlines all output into standardized Pandas DataFrames 
+`Pandas <https://doi.org/10.5281/zenodo.5574486>`_  to allow for easy and quick 
+downstream data analyses (Figure 1). 
+This Pandas-focused setup is ideal if you work with Jupyter notebooks [Kluyver_2016]_.
 
 .. raw:: html
 
@@ -28,45 +29,6 @@ package, a collection of Python modules for structural cheminformatics.
 
 *Figure 1*: OpenCADD-KLIFS fetches KLIFS data offline from a KLIFS download or 
 online from the KLIFS database and formats the output as user-friendly Pandas DataFrames.
-
-The KLIFS database offers a REST API compliant with the OpenAPI specification 
-(`KLIFS OpenAPI <https://dev.klifs.net/swagger_v2/>`_). 
-Our module OpenCADD-KLIFS uses `bravado <https://github.com/Yelp/bravado>`_ to dynamically 
-generate a Python client based on the OpenAPI definitions and adds wrappers to enable the 
-following functionalities:
-
-- A session is set up, which allows access to various KLIFS *data sources* by different 
-  *identifiers* with the API ``session.data_source.by_identifier``. *Data sources* currently 
-  include kinases, structures and annotated conformations, modified residues, pockets, ligands, 
-  drugs, and bioactivities; *identifiers* refer to kinase names, PDB IDs, KLIFS IDs, and more. 
[…]
