### 2021-12-21 #3951 @dominiquesydow
Thank you, @ojeda-e, @andrewtarzia, and @mcs07, for all your input on the manuscript and the documentation!

I have updated the documentation "Statement of need" with the updated manuscript version - and merged PR https://github.com/volkamerlab/opencadd/pull/135.

Will generate the manuscript one more time from the master branch now, to check if everything works after the merge.
