+++ devtools/conda-envs/test_env.yaml
@@ -7,10 +7,10 @@ dependencies:
   # Base depends
   - python
   - pip
-  - MDAnalysis
-  - nglview
   - pandas
   - matplotlib-base
+  - mdanalysis
+  - nglview
   - redo
   # Remove jsonschema once this issue is fixed: https://github.com/Yelp/bravado/issues/478
   - jsonschema<4.0.0
@@ -20,18 +20,20 @@ dependencies:
   - tqdm
   # Chimera reimplementation
   - biotite
-  - biopython 1.77
+  - biopython<=1.77
   # mmligner
   - mmligner
   # theseus
   - muscle
-  - clustalo
   - theseus
   # Testing
   - pytest
   - pytest-cov
   - codecov
-  # docs
+  # Docs
   - sphinx
   - sphinx_rtd_theme
   - nbsphinx
+  # Linting and code style
+  - pylint
+  - black
