+++ devtools/conda-envs/test_env.yaml
@@ -7,24 +7,23 @@ dependencies:
   # Base depends
   - python
   - pip
+  - MDAnalysis
+  - nglview
   - pandas
   - matplotlib-base
-  - mdanalysis
-  - nglview
   - redo
-  # Remove jsonschema once this issue is fixed: https://github.com/Yelp/bravado/issues/478
-  - jsonschema<4.0.0
   - bravado
   - biopandas
   - rdkit
   - tqdm
   # Chimera reimplementation
   - biotite
-  - biopython<=1.77
+  - biopython 1.77
   # mmligner
   - mmligner
   # theseus
   - muscle
+  - clustalo
   - theseus
   # Testing
   - pytest
