### 2022-01-06 #3951 @richardjgowers
@dominiquesydow looks like we're good to go.  Can you make a tagged release of the software (e.g. via zenodo)?  I think it might be easiest to archive the entire package rather than worry about splitting out this module.  (The DOI suggestion seems to be for a poem, top marks for effort though Whedon!)
