### 2022-01-09 #3951 @dominiquesydow
Hi @richardjgowers, thanks!

I have tagged a [new release ](https://github.com/volkamerlab/opencadd/releases/tag/v1.0.1)of the (full) `opencadd` repository - waiting for it to sync with our [zenodo entry](https://doi.org/10.5281/zenodo.5653542). I have contacted zenodo support already and will ping you again once this is solved.

> The DOI suggestion seems to be for a poem, top marks for effort though Whedon!

I love this! :D
