### 2022-01-09 e7e236c
Pin setuptools <60
