### 2022-01-09 fa5f17f
Remove SPHINXOPTS from GHA docs CI
