+++ .github/workflows/ci.yaml
@@ -151,4 +151,4 @@ jobs:
         run: |
           cd docs
           make clean
-          SPHINXOPTS="-W -T --keep-going" make html
+          make html
