### 2022-01-09 d86d652
Add back -T to SPHINXOPTS to GHA docs CI
