### 2022-01-09 c8d834b
Reset SPHINXOPTS!
