+++ .github/workflows/ci.yaml
@@ -151,4 +151,4 @@ jobs:
         run: |
           cd docs
           make clean
-          SPHINXOPTS="-T" make html
+          SPHINXOPTS="-W -T --keep-going" make html
