### 2022-01-10 18719b0
Pin nbsphinx <0.8.8
