### 2022-01-13 volkamerlab/opencadd#139 @dominiquesydow
Zenodo not linked to GitHub releases
Submitted issue: https://github.com/zenodo/zenodo/issues/2281

Important for OpenCADD-KLIFS paper review: https://github.com/openjournals/joss-reviews/issues/3951#issuecomment-1008291863
