+++ opencadd/tests/structure/test_superposition_mmligner.py
@@ -27,13 +27,13 @@ def test_mmligner_alignment():
     assert "metadata" in result
 
     # Check RMSD value for these PDBs
-    assert pytest.approx(result["scores"]["rmsd"], 2.279)
+    assert round(result["scores"]["rmsd"], 3) == 1.469
 
 
 def test_mmligner_alignment_selections():
     aligner = MMLignerAligner()
     structures = [Structure.from_pdbid(pdb_id) for pdb_id in ["4u3y", "4u40"]]
-    user_select = ["backbone and name CA and segid A", "backbone and name CA and segid A"]
+    user_select = ["backbone and name CA and segid B", "backbone and name CA and segid B"]
     selections = [
         structures[0].select_atoms(f"{user_select[0]}"),
         structures[1].select_atoms(f"{user_select[0]}"),
@@ -46,4 +46,4 @@ def test_mmligner_alignment_selections():
     assert "metadata" in result
 
     # Check RMSD value for these PDBs
-    assert pytest.approx(result["scores"]["rmsd"], 2.279)
+    assert round(result["scores"]["rmsd"], 3) == 1.164
