### 2022-02-13 volkamerlab/opencadd#139 @dominiquesydow
Fixed! If ever running into this issue again, try this: https://github.com/zenodo/zenodo/issues/2281#issuecomment-1038361888
