### 2022-02-13 #3951 @dominiquesydow
Hi @richardjgowers,

Good news - the `opencadd` zenodo entry is finally updated to the latest `opencadd` release 🥳
https://zenodo.org/record/6065555

Please let me know if there is anything else for me to do. Thank you!
