### 2022-02-14 #3951 @danielskatz
@dominiquesydow - I've also suggested a few minor changes in the paper in https://github.com/volkamerlab/opencadd/pull/140
