### 2022-02-17 #3951 @dominiquesydow
> @dominiquesydow - I've also suggested a few minor changes in the paper in [volkamerlab/opencadd#140](https://github.com/volkamerlab/opencadd/pull/140)

Thank you, I just merged the suggested changes.
