### 2022-02-17 #3951 @dominiquesydow
> @dominiquesydow - please update the metadata (title, authors) in your zenodo repository so that it matches the paper

I have a quick question regarding this. The zenodo repository covers the whole OpenCADD package, while this paper covers only one module of OpenCADD (i.e., OpenCADD-KLIFS). If I changed the zenodo repo's title and author list to this paper's title and author list, it would not reflect the full content of OpenCADD and the work other people have done for previous versions of OpenCADD anymore.
Would it be possible to not synchronise in this case?
